+++ src/components/Ads/NewAd.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import axios from "axios";
 import styled from "styled-components";
-import Calendar from "../Calendar.js";
-import Uploader from "../Uploader.js";
+import Calendar from "../Calendar.js.js";
+import Uploader from "../Uploader.js.js";
 
 const server = "localhost:5000";
